move into get_messages.rs
