+++ contracts/infinity-pool/src/testing/tests/sim_tests.rs
@@ -1,3 +1,4 @@
+mod get_messages;
 mod helpers;
 mod sim_direct_swap_nfts_for_tokens;
 mod sim_direct_swap_tokens_for_specific_nfts;
